refactor: simplify link-with-providers workflow
