+++ src/core/react-query/release-info/queries.ts
@@ -4,9 +4,10 @@ import { axios } from '@/core/axios';
 
 import type { ReleaseProviderInfoType } from '@/core/react-query/release-info/types';
 
-export const useReleaseInfoProvidersQuery = (noStale = false) =>
+export const useReleaseInfoProvidersQuery = (noStale = false, enabled = true) =>
   useQuery<ReleaseProviderInfoType[]>({
     queryKey: ['release-info', 'providers', noStale],
     queryFn: () => axios.get('ReleaseInfo/Provider'),
     staleTime: noStale ? Infinity : 1000,
+    enabled,
   });
