+++ src/core/utilities/releaseInfoHelpers.ts
@@ -3,7 +3,7 @@ import { produce } from 'immer';
 import { ReleaseSource } from '@/core/types/api/file';
 
 import type { FileType, ReleaseInfoType } from '@/core/types/api/file';
-import type { ManualLinkProviderType, ManualLinkType } from '@/core/types/utilities/unrecognized-utility';
+import type { ManualLinkProviderType, ManualLinkType } from '@/core/types/utilities/link-files-with-providers';
 
 let lastLinkId = 0;
 const generateLinkId = () => {
@@ -47,6 +47,7 @@ const createLinksFromFiles = (files: FileType[], providers: ManualLinkProviderTy
     });
   });
 
+  const hasEnabledProviders = providers.some(provider => provider.enabled);
   const newLinks: Record<number, ManualLinkType> = {};
   sortedFiles.forEach((file) => {
     const now = new Date().toISOString();
@@ -78,7 +79,7 @@ const createLinksFromFiles = (files: FileType[], providers: ManualLinkProviderTy
         id: linkId,
         file,
         providers,
-        state: 'pre-init',
+        state: hasEnabledProviders ? 'searching' : 'init',
         release,
       };
     }
