+++ src/hooks/utilities/useLinkWorkflow.ts
@@ -3,67 +3,28 @@ import { forEach } from 'lodash';
 
 import {
   useAutoPreviewReleaseInfoForFileByIdMutation,
-  usePreviewReleaseInfoByProviderIdMutation,
   useReleaseInfoByFileIdMutation,
   useSubmitReleaseInfoForFileByIdMutation,
 } from '@/core/react-query/release-info/mutations';
 import { mergeReleaseInfo } from '@/core/utilities/releaseInfoHelpers';
 
-import type { ManualLinkType } from '@/core/types/utilities/unrecognized-utility';
+import type { ManualLinkType } from '@/core/types/utilities/link-files-with-providers';
 import type { Updater } from 'use-immer';
 
 const useLinkWorkflow = (
   links: ManualLinkType[],
   setLinks: Updater<Record<number, ManualLinkType>>,
-  providerMap: Record<string, { ID: string, Name: string }>,
   initialized: boolean,
 ) => {
-  const { mutateAsync: previewReleaseInfo } = usePreviewReleaseInfoByProviderIdMutation();
   const { mutateAsync: searchReleaseInfo } = useAutoPreviewReleaseInfoForFileByIdMutation();
   const { mutateAsync: submitReleaseInfo } = useSubmitReleaseInfoForFileByIdMutation();
   const { mutateAsync: fetchReleaseInfo } = useReleaseInfoByFileIdMutation();
 
   // useRef avoids re-renders — these Sets track in-flight API calls (bookkeeping only, not UI state)
   const inFlight = useRef({
-    initializing: new Set<number>(),
     searching: new Set<number>(),
     submitting: new Set<number>(),
     fetching: new Set<number>(),
-  });
-
-  const processPreInit = useEffectEvent((link: ManualLinkType) => {
-    if (inFlight.current.initializing.has(link.id)) return;
-    inFlight.current.initializing.add(link.id);
-
-    const hasProvidersEnabled = link.providers.some(provider => provider.enabled);
-    const offlineImporterProviderId = link.providers.find(
-      provider => providerMap[provider.id]?.Name === 'Offline Importer',
-    )?.id;
-
-    if (!offlineImporterProviderId) {
-      setLinks((draft) => {
-        draft[link.id].state = hasProvidersEnabled ? 'searching' : 'init';
-      });
-      return;
-    }
-
-    const path = link.file.Locations.find(location => location.AbsolutePath)?.AbsolutePath
-      ?? link.file.Locations?.[0]?.RelativePath ?? '';
-
-    previewReleaseInfo({ id: `match://${path}`, providerId: offlineImporterProviderId })
-      .then((data) => {
-        if (!data) return;
-        setLinks((draft) => {
-          draft[link.id].release = data;
-          draft[link.id].state = hasProvidersEnabled ? 'searching' : 'init';
-        });
-      })
-      .catch(() => {
-        setLinks((draft) => {
-          draft[link.id].state = hasProvidersEnabled ? 'searching' : 'init';
-        });
-      })
-      .finally(() => inFlight.current.initializing.delete(link.id));
   });
 
   const processSearch = useEffectEvent((link: ManualLinkType) => {
@@ -73,7 +34,13 @@ const useLinkWorkflow = (
     const enabledReleaseProviders = link.providers
       .filter(provider => provider.enabled)
       .map(provider => provider.id);
-    if (!enabledReleaseProviders.length) return;
+    if (!enabledReleaseProviders.length) {
+      inFlight.current.searching.delete(link.id);
+      setLinks((draft) => {
+        draft[link.id].state = 'init';
+      });
+      return;
+    }
 
     searchReleaseInfo({ fileId: link.file.ID, providerIDs: enabledReleaseProviders })
       .then((data) => {
@@ -145,9 +112,7 @@ const useLinkWorkflow = (
     if (!initialized || !links.length) return;
 
     links.forEach((link) => {
-      if (link.state === 'pre-init') {
-        processPreInit(link);
-      } else if (link.state === 'searching') {
+      if (link.state === 'searching') {
         processSearch(link);
       } else if (link.state === 'submitting') {
         processSubmit(link);
@@ -158,7 +123,6 @@ const useLinkWorkflow = (
   }, [initialized, links]);
 
   useEffect(() => () => {
-    inFlight.current.initializing.clear();
     inFlight.current.searching.clear();
     inFlight.current.submitting.clear();
     inFlight.current.fetching.clear();
